% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dates.R
\docType{data}
\name{available_date_formats}
\alias{available_date_formats}
\title{Available date formats for \code{autoparse_dates} function}
\format{
A character vector of date format strings
}
\usage{
available_date_formats
}
\description{
A character vector containing all supported date formats that can be parsed
by \code{autoparse_dates()}. These formats follow standard date-time components:
\itemize{
\item d: day
\item m: month
\item Y/y: year (Y=4 digits, y=2 digits)
\item B: full month name
\item b: abbreviated month name
\item H: hour
\item M: minute
\item S: second
}
}
\examples{
head(available_date_formats)
}
\keyword{datasets}
