% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_coords.R
\name{big_mark}
\alias{big_mark}
\title{Format Numbers with Thousand Separator}
\usage{
big_mark(x, decimals = NULL, big_mark = ",")
}
\arguments{
\item{x}{A numeric vector to be formatted.}

\item{decimals}{An integer specifying the number of decimal places to
round to. Default is NULL, which means no rounding is performed.}

\item{big_mark}{A character to use as the thousand separator.
Default is ",".}
}
\value{
A character vector of formatted numbers.
}
\description{
This function formats numbers by adding a thousand separator (big mark)
and optionally rounding to a specified number of decimal places.
}
\examples{
big_mark(1234567.89)
big_mark(c(1234.56, 7890123.45), decimals = 2, big_mark = ",")

}
