% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_geonames.R
\name{calculate_string_distance}
\alias{calculate_string_distance}
\title{Calculate String Distances Between Admin Names}
\usage{
calculate_string_distance(admins_to_clean, lookup_admins, method)
}
\arguments{
\item{admins_to_clean}{A vector of administrative names to be cleaned.}

\item{lookup_admins}{A vector of administrative names for lookup.}

\item{method}{The method used to calculate string distances (e.g., "jaro",
"levenshtein"). It can take a number of alo}
}
\value{
A dataframe detailing top N matches for each name, including
algorithm, name, matches, distances, and ranks.
}
\description{
Computes the string distances between administrative names to be cleaned and
a set of lookup administrative names using the specified method. Returns the
top N closest matches for each name.
}
\examples{
# calculate_string_distance(c("New York", "Los Angeles"),
# c("New York", "Los Angeles", "Chicago"),
#   method = "lv"
# )

}
\keyword{internal}
