% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_biannual.R
\name{categorize_biannual}
\alias{categorize_biannual}
\title{Categorize Dates into Biannual Intervals}
\usage{
categorize_biannual(data, date_column)
}
\arguments{
\item{data}{A data frame containing the date column.}

\item{date_column}{A string specifying the name of the date column in the
data frame.}
}
\value{
A data frame with an additional factor column \code{date_categ}
that contains the interval labels.
}
\description{
This function categorizes dates in a given dataset into non-overlapping
six-month intervals, working backward from the maximum date in the dataset.
Each date is assigned to a category based on the interval it falls into.
}
\examples{
data <- data.frame(
  virus_date = as.Date(c("2021-05-01", "2021-11-30", "2022-01-15"))
)
categorized_data <- categorize_biannual(data, "virus_date")

}
