% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_coords.R
\name{check_coord_flip}
\alias{check_coord_flip}
\title{Check for Potentially Flipped Geographical Coordinates}
\usage{
check_coord_flip(
  data,
  shapefile_data = NULL,
  join_key_a,
  join_key_b = NULL,
  lat_col,
  lon_col,
  correct_lat_col = NULL,
  correct_lon_col = NULL
)
}
\arguments{
\item{data}{A dataframe containing the geographical coordinates.}

\item{shapefile_data}{Optional. A dataframe containing reference shapefile
data.}

\item{join_key_a}{The column name in \code{data} to join with \code{shapefile_data}.}

\item{join_key_b}{Optional. The column name in \code{shapefile_data} to join with
\code{data}.}

\item{lat_col}{The name of the latitude column in \code{data}.}

\item{lon_col}{The name of the longitude column in \code{data}.}

\item{correct_lat_col}{Optional. The name of the correct latitude column in
\code{shapefile_data}.}

\item{correct_lon_col}{Optional. The name of the correct longitude column in
\code{shapefile_data}.}
}
\value{
A dataframe with additional columns indicating whether coordinates
are potentially flipped and the distances to the correct coordinates.
}
\description{
This function checks for potentially flipped latitude and longitude
coordinates by comparing them with a reference shapefile, without actually
correcting them.
}
