% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dates.R
\name{check_leap_issue}
\alias{check_leap_issue}
\title{Check Leap Year Issues in Date Column}
\usage{
check_leap_issue(date_col)
}
\arguments{
\item{date_col}{A character vector representing date values.}
}
\value{
A logical vector indicating which dates have leap year issues.
}
\description{
This function checks if the dates in the given column have leap year issues.
It identifies dates that are February 29 in a non-leap year.
}
\examples{
dates <- c("2020-02-29", "2019-02-29", "2021-03-01")
check_leap_issue(dates)
}
