% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ona.R
\name{check_status_api}
\alias{check_status_api}
\title{Check Status of API Response}
\usage{
check_status_api(response)
}
\arguments{
\item{response}{The API response.}
}
\value{
\code{TRUE} for a successful response (status code 200). If the status
code indicates an error (e.g., 400, 403, 500), the function stops
and returns a specific error message.
}
\description{
This function checks the API response status code. It handles various HTTP
status codes by providing specific error messages. It returns \code{TRUE} for
successful responses (status code 200). For error conditions like 400, 403,
500, etc., it stops execution with an error message, aiding in debugging and
issue diagnosis.
}
\examples{
# response <- check_status_api(response)
}
