% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_png_files.R
\name{compress_png_files}
\alias{compress_png_files}
\title{Compress PNG Files in a Directory with pngquant}
\usage{
compress_png_files(directory)
}
\arguments{
\item{directory}{A string specifying the path to the directory containing
PNG files.}
}
\description{
This function iterates over all PNG files in a specified directory
and applies pngquant optimization to each file using the optout package.
}
\examples{
# compress_png_files("path/to/your/folder")

}
