% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_coords.R
\name{correct_flipped_geo_coords}
\alias{correct_flipped_geo_coords}
\title{Correct Flipped Geographical Coordinates}
\usage{
correct_flipped_geo_coords(
  data,
  shapefile_data = NULL,
  join_key_a,
  join_key_b = NULL,
  lat_col,
  lon_col,
  correct_lat_col = NULL,
  correct_lon_col = NULL
)
}
\arguments{
\item{data}{A dataframe containing the geographical coordinates to be
corrected.}

\item{shapefile_data}{Optional. A dataframe containing reference shapefile
data. If NULL, \code{poliprep::shp_global} will be used.}

\item{join_key_a}{The column name in \code{data} to join with \code{shapefile_data}.}

\item{join_key_b}{Optional. The column name in \code{shapefile_data} to join with
\code{data}.
If NULL and \code{shapefile_data} is NULL, defaults to "ADM2_GUID".}

\item{lat_col}{The name of the latitude column in \code{data}.}

\item{lon_col}{The name of the longitude column in \code{data}.}

\item{correct_lat_col}{Optional. The name of the correct latitude column in
\code{shapefile_data}.
If NULL and \code{shapefile_data} is NULL, defaults to "CENTER_LAT".}

\item{correct_lon_col}{Optional. The name of the correct longitude column
in \code{shapefile_data}.
If NULL and \code{shapefile_data} is NULL, defaults to "CENTER_LON".}
}
\value{
A dataframe with corrected latitude and longitude coordinates.
}
\description{
This function corrects potentially flipped latitude and longitude coordinates
by comparing them with a reference shapefile.
}
\examples{
# Using default global shapefile
data <- data.frame(
  id = c(
    "38E294F7-508F-46A2-B5CF-B1D2FD6B2103",
    "1965B8ED-CD56-4784-BEB4-6924A245F533"
  ),
  lat = c(40.7128, -74.0060),
  lon = c(-74.0060, 40.7128)
)

corrected_data <- correct_flipped_geo_coords(
  data,
  join_key_a = "id",
  lat_col = "lat",
  lon_col = "lon"
)

# Using custom shapefile
custom_shapefile <- data.frame(
  region_id = c(
    "38E294F7-508F-46A2-B5CF-B1D2FD6B2103",
    "1965B8ED-CD56-4784-BEB4-6924A245F533"
  ),
  correct_lat = c(40.7128, 41.8781),
  correct_lon = c(-74.0060, -87.6298)
)
corrected_data <- correct_flipped_geo_coords(
  data,
  shapefile_data = custom_shapefile,
  join_key_a = "id",
  join_key_b = "region_id",
  lat_col = "lat",
  lon_col = "lon",
  correct_lat_col = "correct_lat",
  correct_lon_col = "correct_lon"
)

}
