% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{create_gt_table}
\alias{create_gt_table}
\title{Create a GT Table for Data Quality Summary}
\usage{
create_gt_table(
  summary_data,
  title = "Summary of POLIS Data Quality Checks",
  add_nanoplot = TRUE,
  row_start = 6,
  autoscale_nanoplot = FALSE
)
}
\arguments{
\item{summary_data}{A data frame containing the summary of data quality
checks.}

\item{title}{A string for the table title. Default is "Summary of POLIS Data
Quality Checks".}

\item{add_nanoplot}{A logical indicating whether to add a nanoplot column.
Default is TRUE.}

\item{row_start}{An integer indicating the starting row for data color
formatting. Default is 6.}

\item{autoscale_nanoplot}{Logical. Whether to autoscale the nanoplot.
Default is FALSE.}
}
\value{
A GT table object with formatted data quality summary.
}
\description{
This function creates a formatted GT table from a summary of data quality
checks.
}
\details{
The function applies various formatting and styling to the summary data:
\itemize{
\item Applies color scaling to numeric columns based on their values.
\item Sets custom labels for columns.
\item Adds borders to specific rows.
\item Formats numbers to remove decimal places.
\item Optionally adds a nanoplot column for visualizing trends.
}
}
\examples{
#  summary_data <- create_summary_by_group(my_data, "Region", "ID", ...)
# gt_table <- create_gt_table(summary_data,
# title = "My Data Quality Summary",
#  add_nanoplot = TRUE,
#  row_start = 6
# )

}
