% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_date_format.R
\name{detect_date_format}
\alias{detect_date_format}
\title{Detect the Date Format of a Vector of Dates}
\usage{
detect_date_format(date_vector, common_formats = NULL, ...)
}
\arguments{
\item{date_vector}{A character vector of dates.}

\item{common_formats}{An optional character vector specifying the date
formats to try. If NULL, a default set of formats is used.}

\item{...}{Additional arguments passed to \code{lubridate::parse_date_time}.}
}
\value{
A character string representing the date format that had the highest
success rate of parsing the dates in \code{date_vector}. Returns
\code{NA} if no suitable format is found.
}
\description{
This function automatically detects and returns the most likely date format
from a given vector of date strings. It attempts to parse the dates using
a set of common date formats and selects the format that successfully parses
the majority of the entries.
}
\examples{
# date_vector <- c("01-02-2021", "03-15-1988", "2023-05-12",
#                  "19/02/2021", "1988-15-03", "05-12-2023", "05.27.2022")
# detect_date_format(date_vector)

}
