% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_and_read_mdb.R
\name{export_and_read_mdb}
\alias{export_and_read_mdb}
\title{Export and Read Microsoft Access Database Table}
\usage{
export_and_read_mdb(mdb_path, table_name, output_csv = "/tmp/temp_export.csv")
}
\arguments{
\item{mdb_path}{Character string. Full path to the .mdb file.}

\item{table_name}{Character string. Name of the table to export from the
.mdb file.}

\item{output_csv}{Character string. Path for the temporary CSV file.
Default is "/tmp/temp_export.csv".}
}
\value{
A dataframe containing the data from the specified Access table.
}
\description{
This function exports a specified table from a Microsoft Access database
(.mdb file) to a CSV file, then reads the CSV into an R dataframe.
}
\details{
This function requires 'mdb-tools' to be installed on the system.
If not installed, it will attempt to install it automatically on macOS
(using Homebrew) and Linux (using apt-get). For Windows, manual installation
is required. The function uses the system command 'mdb-export' to export
the table to CSV, mdb-schema' to get column information, and '
poliprep::read' for importing.
}
\examples{
\dontrun{
df <- export_and_read_mdb(
    "/path/to/database.mdb",
    "TableName"
)
}

}
