% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polis.R
\name{extract_entity_sets}
\alias{extract_entity_sets}
\title{Extract Entity Sets from the POLIS API Metadata}
\usage{
extract_entity_sets(url)
}
\arguments{
\item{url}{A character string specifying the URL of the POLIS API metadata.}
}
\value{
A data frame containing two columns:
\describe{
\item{DataName}{The name of the data set.}
\item{EntityType}{The entity type associated with the data set.}
}
}
\description{
This function retrieves and extracts entity sets from the POLIS API metadata
URL provided. It parses the XML content to extract the names and entity
types of the available data sets.
}
\examples{
\dontrun{
# Define the URL for the POLIS API metadata
polis_api_root_url <-
             "https://extranet.who.int/polis/api/v2/$metadata?token="
api_token <- Sys.getenv("POLIS_API_KEY")
url <- paste0(polis_api_root_url, api_token)

# Extract the entity sets
xml_df <- extract_entity_sets(url)
}
}
