% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_geonames.R
\name{format_choice}
\alias{format_choice}
\title{Format a single choice for display}
\usage{
format_choice(index, choice, width)
}
\arguments{
\item{index}{Numeric index of the choice}

\item{choice}{Character string of the choice text}

\item{width}{Total width of the formatted choice}
}
\value{
Formatted choice string
}
\description{
Format a single choice for display
}
