% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_geonames.R
\name{format_choices}
\alias{format_choices}
\title{Format all choices into columns}
\usage{
format_choices(choices, num_columns, column_width = 45)
}
\arguments{
\item{choices}{Vector of choice strings}

\item{num_columns}{Number of columns to display}

\item{column_width}{Width of each column}
}
\value{
Vector of formatted choice strings, one per row
}
\description{
Format all choices into columns
}
