% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_detections.R
\name{format_date_ord}
\alias{format_date_ord}
\title{Format Date with Ordinal Day}
\usage{
format_date_ord(date_column)
}
\arguments{
\item{date_column}{A Date vector or a character vector that can be coerced
to Date. The dates to be formatted.}
}
\value{
A character vector representing the formatted dates with ordinal day,
abbreviated month, and year (e.g., "2nd Jan 2020").
}
\description{
This function converts a date column to a string with an ordinal day
(e.g., "1st", "2nd", "3rd") followed by the abbreviated month name and year.
}
\details{
The function handles special cases for 11th, 12th, and 13th, which always
use "th" as the ordinal suffix regardless of their ones digit.
}
\examples{
format_date_ord(as.Date("2020-01-02")) # Returns "2nd Jan 2020"
format_date_ord(c("2023-03-21", "2023-04-11"))
# Returns c("21st Mar 2023", "11th Apr 2023")

}
\seealso{
\code{\link{as.Date}}, \code{\link{format}}
}
