% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ona.R
\name{generate_urls}
\alias{generate_urls}
\title{Generate URLs for Data Retrieval}
\usage{
generate_urls(full_data, file_path, data_file_name, base_url, form_ids)
}
\arguments{
\item{full_data}{The full dataset containing form IDs and last edited dates.}

\item{file_path}{The path to the directory containing the data file.}

\item{data_file_name}{The name of the data file (without extension).}

\item{base_url}{The base URL of the API endpoint.}

\item{form_ids}{A vector of form IDs for which URLs need to be generated.}
}
\value{
A list of URLs corresponding to the provided form IDs.
}
\description{
This function generates a list of URLs based on a given set of form IDs.
If data exists for a form ID, the URL will include the last edited date
parameters; otherwise, the URL will only contain the form ID.
}
\examples{
\dontrun{
generate_urls(file_path = "/path/to/data", 
              data_file_name = "data_file", 
              base_url = "https://example.com", 
              form_ids = c(8417, 8734, 9056))
}

}
