% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polis.R
\name{get_api_date_suffix}
\alias{get_api_date_suffix}
\title{Get Date Field and Endpoint Suffix Based on Data Type}
\usage{
get_api_date_suffix(data_type)
}
\arguments{
\item{data_type}{A string specifying the type of data for which information
is needed. Valid data types include 'cases', 'virus',
'population', 'env', 'geo', 'geo_synonym', 'im', 'activity',
'lab_specimen', 'lab_specimen_virus', 'lab_env',
'sub_activ', and lqas'.}
}
\value{
A list with two elements: 'endpoint_suffix', which is the suffix for
the API endpoint corresponding to the data type, and 'date_field',
which is the name of the date field relevant to the data type.
}
\description{
This function returns the appropriate endpoint suffix and date field name
for a given data type. It is specifically designed to work with the POLIS API
data retrieval system. The function takes a data type as input and returns a
list containing the corresponding endpoint suffix and date field. It ensures
that the data type provided is valid and returns an error message if not.
}
\examples{
result <- get_api_date_suffix("cases")
endpoint_suffix <- result$endpoint_suffix
date_field <- result$date_field
}
