% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_detections.R
\name{get_detections}
\alias{get_detections}
\title{Get virus detections from data}
\usage{
get_detections(data, virus_types)
}
\arguments{
\item{data}{A data frame containing virus detection data}

\item{virus_types}{A vector of virus types to filter}
}
\value{
A character vector of detection counts by virus type
}
\description{
Get virus detections from data
}
