% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{get_missing_ids}
\alias{get_missing_ids}
\title{Get IDs with Missing Values for a Specific Column}
\usage{
get_missing_ids(data, id_column, variable)
}
\arguments{
\item{data}{A data frame to check for missing values.}

\item{id_column}{A string specifying the name of the ID column.}

\item{variable}{A string specifying the name of the column to
check for missing values.}
}
\value{
A vector of IDs corresponding to rows with missing values in
the specified column.
}
\description{
This function returns the IDs of rows where a specified column has
missing values.
}
\examples{
df <- data.frame(
  id = 1:5,
  a = c(1, NA, 3, NA, 5),
  b = c("x", "y", NA, "w", "z")
)
get_missing_ids(df, "id", "a")
get_missing_ids(df, "id", "b")

}
