% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ona.R
\name{get_ona_data}
\alias{get_ona_data}
\title{Get Data from ONA API}
\usage{
get_ona_data(
  base_url = "https://api.whonghub.org",
  form_id,
  api_token,
  selected_columns = NULL,
  logical_filters = NULL,
  comparison_filters = NULL
)
}
\arguments{
\item{base_url}{A string specifying the URL for the ONA API.
Default is "https://api.whonghub.org"}

\item{form_id}{A string or numeric value specifying the dataset ID.
Can be found in the URL of the form on ONA.}

\item{api_token}{A string specifying the API token for authentication}

\item{selected_columns}{Optional. A character vector of column names to
download. Default is NULL which downloads all columns.
Example: c("year", "form_id")}

\item{logical_filters}{Optional. A named list for filtering with logical
operators. Names are column names, values are vectors to
filter by. Uses OR within groups and AND between groups.}

\item{comparison_filters}{Optional. A named list for filtering with
comparisons. Names are column names, values are comparison
conditions. Supports >, <, >=, <=, =, != operators.}
}
\value{
A data frame containing the filtered data downloaded from the ONA
API, with empty columns removed.
}
\description{
This function gets data for a specified form from the ONA API using a
provided API token and form ID. It returns the data in a structured format
if the request is successful. If the request fails, the function stops and
returns an error message indicating the failure reason.
}
\examples{
\dontrun{
# Basic usage
data <- get_ona_data(
  form_id = "123456",
  api_token = "your_api_token_here"
)

# Basic usage with selected columns and filters
ona_eim_ipds_ih_epm <- get_ona_data(
  form_id = 7178,
  api_token = ONA_TOKEN,
  selected_columns = c("states", "endtime", "today", "_duration"),
  logical_filters = list(states = c("BORNO", "KANO")),
  comparison_filters = (~ `today` >= "2023-02-04" & `today` <= "2025-02-04")
)
}

}
\seealso{
\url{https://api.ona.io/api/v1/data/} for ONA API documentation
}
