% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polis.R
\name{get_polis_api_data}
\alias{get_polis_api_data}
\title{Retrieve Data from POLIS API}
\usage{
get_polis_api_data(
  min_date = "2021-01-01",
  max_date = Sys.Date(),
  data_type = "cases",
  region = "AFRO",
  country_code = NULL,
  select_vars = NULL,
  updated_dates = FALSE,
  polis_api_key,
  log_results = FALSE,
  log_file_path
)
}
\arguments{
\item{min_date}{Start date for data retrieval in 'YYYY-MM-DD' format.
Specifies the earliest date of the data to be fetched.
Required parameter.}

\item{max_date}{End date for data retrieval in 'YYYY-MM-DD' format.
Defaults to the current system date. Specifies the latest
date of the data to be fetched.}

\item{data_type}{Type of data to retrieve.
Supported types include 'cases', 'virus', 'population',
'env' (Environmental), 'geo' (Geographical), 'geo_synonym',
'im' (Independent Monitoring), 'activity', 'sub_activ'
(Sub-activities), and 'lqas' (Lot Quality Assurance
Sampling), 'lab_specimen' (Human Specimen),
'lab_env' (Environmental Sample), 'lab_specimen_virus'
(Human Specimen Viruses). Default is 'cases'.}

\item{region}{Region code for data filtering.
Represents the WHO region from which to retrieve the data.
Possible values are AFRO; AMRO; EMRO; EURO; SEARO; WPRO Use
'Global' to retrieve global data. Default is 'AFRO'.}

\item{country_code}{ISO3 country code to filter the data. Default is NULL.}

\item{select_vars}{Vector of variables to select from the API response.
If NULL (default), all variables are selected.}

\item{updated_dates}{Logical indicating whether to use the 'LastUpdateDate'}

\item{polis_api_key}{API key for authentication.
Default is retrieved from the environment variable
'POLIS_API_KEY'. An explicit API key can be provided
if required.}

\item{log_results}{Logical indicating whether to log the results of the API}

\item{log_file_path}{Path to the directory where the log file will be saved}
}
\value{
A data frame containing the requested data aggregated from all pages
of the API response. Each row represents a record, and columns
correspond to the variables in the dataset.
}
\description{
This function serves as a gateway to retrieve various types of health-related
data from the POLIS API. It simplifies the process of querying the API by
handling authentication, constructing the request URL, and iterating over
paginated results. The function is versatile, allowing for data retrieval
based on a range of parameters such as date range, data type, region, and
specific variables. It automates the handling of API responses, including
status checking and data aggregation, and returns the results in a
convenient data frame format. The function is designed to be robust,
providing informative error messages in case of missing API keys or
unsuccessful API calls.
}
\examples{
\dontrun{
data <- get_polis_api_data("2021-01-01", "2021-01-31", "cases", "AFRO")
}
}
