% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{get_polis_detections}
\alias{get_polis_detections}
\title{Get detections from POLIS data}
\usage{
get_polis_detections(data, case_type = c("ES", "AFP"))
}
\arguments{
\item{data}{A POLIS dataset}

\item{case_type}{Type of cases to filter ("ES" or "AFP")}
}
\value{
Filtered dataset containing only specified case type
}
\description{
Get detections from POLIS data
}
