% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polis.R
\name{get_status_code}
\alias{get_status_code}
\title{Get HTTP Status Code for a POLIS API Table}
\usage{
get_status_code(table, api_token = Sys.getenv("POLIS_API_KEY"))
}
\arguments{
\item{table}{A character string specifying the name of the table to check.}

\item{api_token}{A character string of the POLIS API token. Defaults to
\code{Sys.getenv("POLIS_API_KEY")}.}
}
\value{
An integer representing the HTTP status code, or a character string
("Timeout" or "Error") if an exception occurs.
}
\description{
This function sends a GET request to a specified POLIS API table and returns
the HTTP status code. It includes error handling to manage timeouts and other
potential request errors.
}
\examples{
\dontrun{
# Get the status code for the "countries" table
status_code <- get_status_code("countries")
}
}
