% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ona.R
\name{get_updated_ona_data}
\alias{get_updated_ona_data}
\title{Update ONA Data}
\usage{
get_updated_ona_data(
  base_url = "https://api.whonghub.org",
  form_ids,
  api_token,
  log_results = TRUE,
  file_path = NULL,
  data_file_name = "my_ona_data"
)
}
\arguments{
\item{base_url}{The base URL for the API calls, defaults to
"https://api.whonghub.org".}

\item{form_ids}{A vector of form IDs to download data for.}

\item{api_token}{The API token used for authentication with the API.}

\item{log_results}{Boolean flag to indicate whether to log results of the
data updates, defaults to TRUE.}

\item{file_path}{The file path where data files will be stored, can be NULL.}

\item{data_file_name}{The base name for the data file, defaults to
"my_ona_data".}
}
\value{
Returns a data frame containing the combined new and existing data.
}
\description{
This function manages and updates a dataset by downloading new data from
specified form IDs using an API endpoint, and integrating this with existing
data. The function checks for the availability of form IDs, downloads new
data, combines it with existing data if present, logs the results if
specified, and saves the updated dataset.
}
\examples{
# get_updated_ona_data(
#       form_ids = c(123, 456), api_token = "your_api_token_here")

}
