% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dictionary.R
\name{get_var_labels}
\alias{get_var_labels}
\title{Get Variable Labels from var_tree.yml}
\usage{
get_var_labels(language = "en")
}
\arguments{
\item{language}{Language code ("en" or "fr"). Default is "en".}
}
\value{
Named character vector mapping variable names to labels.
}
\description{
Load variable labels from the package's var_tree.yml file.
}
\examples{
labels <- get_var_labels()
labels["country_iso3"]

}
