% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_geonames.R
\name{handle_file_save}
\alias{handle_file_save}
\title{Save Dataframe to RDS File}
\usage{
handle_file_save(data_to_save, default_save_path = NULL)
}
\arguments{
\item{data_to_save}{DataFrame to be saved.}

\item{default_save_path}{Default path for saving the dataframe. If not
provided or invalid, the user is prompted for a new path.
Defaults to \code{NULL}.}
}
\value{
Invisible \code{NULL}. The function's primary purpose is to saving a
file, not to return a value.
}
\description{
Prompts the user for confirmation before saving a dataframe to a RDS file.
If the specified file path does not exist or is NULL, the user is prompted to
provide a new path. A default file name is used if no valid path is provided.
If a cache file already exists, the function merges new data with existing
data, ensuring that the most recent entries (based on \code{created_time})
are retained.
}
\examples{
# handle_file_save(data_to_save, "path/to/default/location.rds")

}
\keyword{internal}
