% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_geonames.R
\name{handle_user_interaction}
\alias{handle_user_interaction}
\title{Interact with Users for Data Cleaning Choices}
\usage{
handle_user_interaction(
  input_data,
  levels,
  level,
  clear_console = TRUE,
  stratify,
  max_options
)
}
\arguments{
\item{input_data}{Data frame containing admin names etc.,.}

\item{levels}{The avaiable admin levels withnin the prep_geoname function}

\item{level}{The admins level being cleaned, i.ie level1 or even disrict.}

\item{clear_console}{Logical, whether to clear the console before showing
prompts; defaults to TRUE.}

\item{stratify}{Logical, whether to stratify the cleaning process.}

\item{max_options}{Maximum number of options to display in the menu.
Default is 200.}
}
\value{
Data frame of user-selected replacements if any; otherwise,
provides feedback based on user actions.
}
\description{
Presents an interactive CLI menu using \code{cli} for users to make selections on
data cleaning choices, particularly for administrative names like countries,
provinces, districts, and subdistricts. It allows users to replace, skip,
save, or exit, incorporating user feedback into the data cleaning process.
}
\examples{
# handle_user_interaction(my_data, "level1", TRUE, TRUE, 20)

}
\keyword{internal}
