% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polis.R
\name{iterative_api_call}
\alias{iterative_api_call}
\title{Try API Call with Retry Logic}
\usage{
iterative_api_call(url, token = NULL, max_attempts = 3)
}
\arguments{
\item{url}{The URL for the API endpoint to which the GET request is made.}

\item{token}{The authorization token required for the API request.}

\item{max_attempts}{The maximum number of attempts to make the API call.
Default is 3 attempts.}
}
\value{
The response from the API call if successful, or an error message
if all attempts fail.
}
\description{
This function attempts to make an API call to a specified URL using the GET
method. It includes a retry mechanism, where the request is attempted up to a
specified number of times (\code{max_attempts}) before giving up. If the request
is successful and returns a 200 status code, the response is returned. If any
errors occur or if a non-200 status code is received, the function will wait
for 5 seconds and retry the request. After the maximum number of attempts, if
the request still fails, the function stops with an error message.
}
\examples{
\dontrun{
iterative_api_call("https://api.example.com/data", "your_api_token")
}
}
