% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{join_and_check_mismatches}
\alias{join_and_check_mismatches}
\title{Check for Mismatches Between Data and Shapefile}
\usage{
join_and_check_mismatches(
  data,
  shapefile_data = NULL,
  geo_name_cols,
  geo_id_cols,
  shapefile_name_cols = NULL,
  shapefile_id_cols = NULL,
  id_col
)
}
\arguments{
\item{data}{A data frame containing the geographical data to be checked.}

\item{shapefile_data}{A data frame containing the shapefile data. If NULL,
poliprep::shp_global will be used.}

\item{geo_name_cols}{A character vector of column names in 'data' containing
geographical names.}

\item{geo_id_cols}{A character vector of column names in 'data' containing
geographical IDs. The first element (typically ADM0_GUID) will be
removed.}

\item{shapefile_name_cols}{A character vector of column names in
'shapefile_data' containing geographical names. If NULL, default
values will be used.}

\item{shapefile_id_cols}{A character vector of column names in
'shapefile_data' containing geographical IDs. If NULL, default values
will be used.}

\item{id_col}{A string specifying the name of the ID column in 'data'.}
}
\value{
A list containing two data frames:
\item{summary_table}{A summary of mismatches for each geographical column}
\item{detailed_mismatches}{Detailed information about each mismatch}
}
\description{
This function compares geographical data with a shapefile to identify
mismatches in names and IDs. It provides both a summary of mismatches and
detailed information about specific mismatches.
}
\examples{
# Sample data
sample_data <- data.frame(
  EPID = 1:5,
  Admin0GUID = c("A1", "A1", "A2", "A2", "A3"),
  Admin0Name = c("Country1", "Country1", "Country2", "Country2", "Country3"),
  Admin1GUID = c("B1", "B2", "B3", "B3", "B4"),
  Admin1Name = c(
    "Province1", "Province2", "Province3", "Province3",
    "Province4"
  ),
  Admin2GUID = c("C1", "C2", "C3", "C4", "C5"),
  Admin2Name = c(
    "District1", "District2", "District3", "District4",
    "District5"
  )
)

# Sample shapefile data
shapefile_data <- data.frame(
  ADM0_NAME = c("Country1", "Country2"),
  ADM1_NAME = c("Province1", "Province3"),
  ADM2_NAME = c("District1", "District3"),
  ADM1_GUID = c("B1", "B3"),
  ADM2_GUID = c("C1", "C3")
)

# Define column names
geo_name_cols <- c("Admin0Name", "Admin1Name", "Admin2Name")
geo_id_cols <- c("Admin0GUID", "Admin1GUID", "Admin2GUID")
shapefile_name_cols <- c("ADM0_NAME", "ADM1_NAME", "ADM2_NAME")
shapefile_id_cols <- c("ADM1_GUID", "ADM2_GUID")
id_col <- "EPID"

# Run the function
results <- join_and_check_mismatches(
  sample_data, shapefile_data,
  geo_name_cols, geo_id_cols,
  shapefile_name_cols, shapefile_id_cols,
  id_col
)

# View results
print(results$summary_table)
print(results$detailed_mismatches)
}
