% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lakes}
\alias{lakes}
\title{World lakes data}
\format{
An sf object
}
\source{
Natural Earth \url{https://www.naturalearthdata.com/}
}
\usage{
lakes
}
\description{
A dataset containing high-resolution world lakes data.
}
\keyword{datasets}
