% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ona.R
\name{praise_emoji}
\alias{praise_emoji}
\title{Generate a Random Emoji for Successful Results}
\usage{
praise_emoji()
}
\value{
A single emoji character if UTF-8 output is supported; otherwise,
an empty string.
}
\description{
Selects a random emoji from a predefined list to use as a celebratory message
in user interfaces or logs where UTF-8 output is supported.
Taken from testthat package.
}
\examples{
# emoji <- praise_emoji()

}
