% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_match_names.R
\name{prep_match_names}
\alias{prep_match_names}
\title{Match naming conventions}
\usage{
prep_match_names(ref_dataframe, target_dataframe)
}
\arguments{
\item{ref_dataframe}{reference dataframe to be matched to}

\item{target_dataframe}{target dataframe that needs to be matched to
ref_dataframe}
}
\value{
returns the target dataframe with common names matching the reference
dataframe
}
\description{
Given two dataframes with variables in common but with diffrent naming
conventions (ex: CamelCase vs camelcase), this function will find the common
varibles and rename those in the target table with those from the reference
table.
}
\examples{

# Create reference and target dataframes
ref_dataframe <- tibble::tibble(CountryName = c("Cameroon",
                                                "Nigeria",
                                               "Tchad",
                                               "Niger"),
                               NameOfCapital = c("Yaounde",
                                                 "Abuja",
                                                 "Ndjamena",
                                                 "Niamey"))

target_dataframe <- tibble::tibble(countryname = c("South Sudan",
                                                   "Kenya",
                                                   "Ethiopia",
                                                   "CAR"),
                                   nameofcapital = c("Juba",
                                                     "Nairobi",
                                                     "Addis Ababa",
                                                     "Bangui"))
# Check the names matching. Expect False
names(ref_dataframe) == names(target_dataframe)

# Match the dataframes names
target_dataframe <- prep_match_names(ref_dataframe, target_dataframe)

# Check the names matching. Expect True
names(ref_dataframe) == names(target_dataframe)
}
