% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_detections.R
\name{prep_new_detections_table}
\alias{prep_new_detections_table}
\title{Prepare New Detections Table}
\usage{
prep_new_detections_table(
  polis_df_old,
  polis_df_new,
  polis_sia,
  include_sia = TRUE,
  save_table = FALSE,
  output_path = NULL,
  vheight = 800,
  vwidth = 3260
)
}
\arguments{
\item{polis_df_old}{A data frame containing the old POLIS data.}

\item{polis_df_new}{A data frame containing the new POLIS data.}

\item{polis_sia}{A data frame containing SIA data.}

\item{include_sia}{Logical, whether to include SIA information in the output
table. Default is TRUE.}

\item{save_table}{Logical, whether to save the table as an image file.
Default is FALSE.}

\item{output_path}{Character string specifying the directory to save the
output file. If NULL, uses the current working directory.}

\item{vheight}{Numeric, the height of the output image in pixels. Default is
800.}

\item{vwidth}{Numeric, the width of the output image in pixels. Default is
3260.}
}
\value{
A gt table object containing the summary of poliovirus detections.
}
\description{
This function creates a summary table of poliovirus detections, comparing old
and new data, and optionally includes information about Supplementary
Immunization Activities (SIA).
}
\details{
The function performs the following main steps:
\enumerate{
\item Checks for required packages and loads them if necessary.
\item Processes the old POLIS data to get previous detection dates.
\item Filters and processes the new POLIS data.
\item Combines and summarizes the data, including calculations for various
metrics.
\item Joins with SIA data if include_sia is TRUE.
\item Creates a formatted gt table with conditional styling.
\item Optionally saves the table as an HTML and PNG file.
}
}
\note{
\itemize{
\item The function uses various dplyr, tidyr, and gt functions for data
manipulation and table creation.
\item It applies conditional formatting to highlight new emergences, delayed
reporting, and other important information.
\item If save_table is TRUE, it saves the table as both HTML and PNG files, then
deletes the HTML file.
}
}
\examples{
# Basic usage:
# result_table <- prep_new_detections_table(
#   polis_df_old, polis_df_new,
#   polis_sia
# )

}
