% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ona_data_endpoints.R
\name{prep_ona_data_endpoints}
\alias{prep_ona_data_endpoints}
\title{Get Data endpoints from ONA API}
\usage{
prep_ona_data_endpoints(base_url = "https://api.whonghub.org", api_token)
}
\arguments{
\item{base_url}{A string specifying the base URL for the ONA API,
could be https://esurv.afro.who.int eSurv.
Default is https://api.whonghub.org}

\item{api_token}{A string specifying the API token for ONA}
}
\value{
A list containing the data geted from the ONA API.
}
\description{
This function getes data endpoints for a specified form from the ONA API
using a provided API token and form ID. It returns the data in a structured
format if the request is successful. If the request fails, the function stops
and returns an error message indicating the failure reason.
}
\examples{
# base_url <- https://api.ona.io
# api_token <- "your_api_token_here"
# data <- get_ona_data(base_url, api_token)
}
\seealso{
\url{https://api.ona.io/api/v1/data/} for more info on ONA API
}
