% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polis.R
\name{process_api_response}
\alias{process_api_response}
\title{Process API Response}
\usage{
process_api_response(response)
}
\arguments{
\item{response}{The response object from an API call made using the \code{httr2}
package.}
}
\value{
A \code{dataframe}, which is the parsed data from the response If the
response is unsuccessful, both elements of the list are NULL.
}
\description{
This function processes the response from an API call. It checks the status
code of the response; if the status code is 200, indicating a successful
response, it parses the content of the response using.
}
\examples{
\dontrun{
# Example usage within a function that makes API calls:
response <- iterative_api_call(api_url, token)
processed_response <- process_response(response)
}
}
