% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{Read in Data and Shapefiles from Various File Formats}
\usage{
read(file_path, ...)
}
\arguments{
\item{file_path}{Character string specifying the path to the input file or
a URL pointing to the dataset.}

\item{...}{Additional arguments to be passed to the underlying read
functions. These arguments are specific to the file format being imported.
Please refer to the documentation of each package used for more
information.}
}
\value{
A data frame or appropriate R object containing the imported data.
}
\description{
This function provides a unified interface for reading data from various
file formats supported by the \code{\link[=rio]{rio}} and
\code{\link[=sf]{sf}} package. The format is automatically detected from the
file extension to simplify the importing process.
}
\examples{
# Locate test data directory
path <-  system.file("extdata",
                     package = "poliprep")

# Import a CSV file
data_csv <- read(file_path = file.path(path, "test_data.csv"))

# Import an Excel file
data_excel <- read(file_path = file.path(path, "test_data.xlsx"))

# Import a Stata DTA file
data_dta <- read(file_path = file.path(path, "test_data.dta"))

# Import an RDS file
data_rds <- read(file_path = file.path(path, "test_data.rds"))

# Import an RData file
data_rdata <- read(file_path = file.path(path, "test_data.RData"))

# Import an SPSS file
data_spss <- read(file_path = file.path(path, "test_data.sav"))

# Import an shapefiles file (GeoJSON/json)
data_spss <- read(file_path = file.path(path, "test_data.GeoJSON"))

}
\seealso{
\code{\link[=rio]{rio::import()}} and
\code{\link[=sf]{sf::read_sf()}},  which this function is based on.
}
