% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save}
\alias{save}
\title{Save Data and Shapefiles to Various File Formats}
\usage{
save(data, file_path, ...)
}
\arguments{
\item{data}{The dataset to be saved}

\item{file_path}{Character string specifying the path to the output file.}

\item{...}{Additional arguments to be passed to the underlying write
functions. These arguments are specific to the file format being saved.
Please refer to the documentation of each package used for more
information.}
}
\value{
No return value, called for side effects.
}
\description{
This function provides a unified interface for saving data to various
file formats supported by the \code{\link[=rio]{rio::export()}}
function. The format is automatically detected from the file extension to
simplify the saving process.
}
\examples{
# Create temporary account
tmpdir <- tempfile()
dir.create(tmpdir)

# Save a CSV file
save(mtcars, file_path = file.path(tmpdir, "file.csv"))

# Save an Excel file
save(mtcars, file_path = file.path(tmpdir, "file.xlsx"))

# Save a Stata DTA file
save(mtcars, file_path = file.path(tmpdir, "file.dta"))

# Save an RDS file
save(mtcars, file_path = file.path(tmpdir, "file.rds"))

# Save an RData file
save(list(mtcars = mtcars, iris = iris),
       file_path = file.path(tmpdir, "file.RData"))

# For saving shapefiles 
# make example shape data
my_shp <-  sf::st_sfc(
  sf::st_point(c(43, 23))) |> 
    cbind(mtcars[1, ]) |> 
    sf::st_as_sf(crs = sf::st_crs(4326))

# save an RDS file
save(my_shp, file_path = file.path(tmpdir, "file.shp"))

# Remove the temporary directory and its contents
unlink(tmpdir, recursive = TRUE)
}
\seealso{
\code{\link[=rio]{rio::import()}} and
\code{\link[=sf]{sf::read_sf()}},  which this function is based on.
}
