% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polis.R
\name{save_polis_data}
\alias{save_polis_data}
\title{Save Polis Data to compressed RDS File}
\usage{
save_polis_data(polis_data, polis_path, filname, max_datasets = 5)
}
\arguments{
\item{polis_data}{The POLIS data object to be saved.}

\item{polis_path}{The directory path where the RDS file will be saved. This
function will check this directory for existing datasets and will maintain
only the 5 most recent datasets, deleting older ones.}

\item{filname}{The name of the file to be saved.}

\item{max_datasets}{The max number of datasets to retain in the directory.}
}
\value{
Invisible NULL. This function is used for its side effect of
saving a file and potentially deleting older files, rather than for
returning a value.
}
\description{
This function saves a given POLIS data object to an RDS file using a
specific naming convention based on the current date. It also manages the
retention of only the 5 most recent datasets in the specified directory,
removing older datasets if necessary.
}
\examples{
# Assume `polis_data` is your dataset and `./polis_datasets` is your
# target directory
# save_polis_data(polis_data, "./polis_datasets")

}
