% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_coords.R
\name{setup_coord_operation}
\alias{setup_coord_operation}
\title{Common Setup for Coordinate Operations}
\usage{
setup_coord_operation(
  data,
  shapefile_data = NULL,
  join_key_a,
  join_key_b = NULL,
  lat_col,
  lon_col,
  correct_lat_col = NULL,
  correct_lon_col = NULL
)
}
\arguments{
\item{data}{A dataframe containing the geographical coordinates.}

\item{shapefile_data}{Optional. A dataframe containing reference shapefile
data.}

\item{join_key_a}{The column name in \code{data} to join with \code{shapefile_data}.}

\item{join_key_b}{Optional. The column name in \code{shapefile_data} to join with
\code{data}.}

\item{lat_col}{The name of the latitude column in \code{data}.}

\item{lon_col}{The name of the longitude column in \code{data}.}

\item{correct_lat_col}{Optional. The name of the correct latitude column in
\code{shapefile_data}.}

\item{correct_lon_col}{Optional. The name of the correct longitude column in
\code{shapefile_data}.}
}
\value{
A list containing processed data and shapefile data.
}
\description{
This helper function performs common setup tasks for coordinate operations.
}
