% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dates.R
\name{try_parsing}
\alias{try_parsing}
\title{Try parsing dates with multiple formats}
\usage{
try_parsing(x, formats)
}
\arguments{
\item{x}{A character vector of dates to parse}

\item{formats}{A character vector of date formats to try (see ?strptime for
format specifications)}
}
\value{
A POSIXct vector of parsed dates
}
\description{
This helper function attempts to parse dates using specified formats and
handles ISO 8601 formatted dates (containing 'T') separately using ymd_hms.
}
