% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ona.R
\name{update_ona_data}
\alias{update_ona_data}
\title{Update Data from ONA API}
\usage{
update_ona_data(
  base_url = "https://api.whonghub.org",
  form_ids,
  api_token,
  log_results = TRUE,
  file_path = NULL,
  data_file_name = "my_ona_data"
)
}
\arguments{
\item{base_url}{Base URL for the ONA API endpoint, defaulting to a predefined
URL.}

\item{form_ids}{The ID(s) of the form to download data for.}

\item{api_token}{Authorization token for accessing the API.
with new data from ONA.}

\item{log_results}{Logical indicating whether to log the results of the data
update.}

\item{file_path}{Optional path to save the downloaded data; defaults to the
working directory.}

\item{data_file_name}{Name of the file to save the data into, defaults to
my_ona_data".}
}
\value{
Returns a data frame containing the retrieved and possibly updated
data.
}
\description{
Retrieves and optionally updates data from the ONA API for a specified form.
It checks the availability of the form ID, downloads new data if requested,
and can log results. It also handles pagination of data automatically.
}
\examples{
# update_data <- update_ona_data(
# base_url = "https://api.whonghub.org",
#   form_id = "12345",
#   api_token = "your_api_token_here",
#   update_ona = TRUE,
#   log_results = TRUE
#)

}
