% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polis.R
\name{update_polis_api_data}
\alias{update_polis_api_data}
\title{Retrieve and Update POLIS Data}
\usage{
update_polis_api_data(
  min_date,
  max_date = Sys.Date(),
  data_type = "cases",
  region = "AFRO",
  country_code = NULL,
  select_vars = NULL,
  file_path = NULL,
  save_directly = FALSE,
  log_results = FALSE,
  polis_api_key = NULL
)
}
\arguments{
\item{min_date}{Start date for data retrieval in 'YYYY-MM-DD' format.
This date marks the beginning of the dataset to be retrieved.}

\item{max_date}{End date for data retrieval, default is the current date.
This date marks the end of the dataset to be retrieved.}

\item{data_type}{Type of data to retrieve.
Supported types include 'cases', 'virus', 'population',
'env' (Environmental), 'geo' (Geographical), 'geo_synonym',
'im' (Independent Monitoring), 'activity', 'sub_activ'
(Sub-activities), and 'lqas' (Lot Quality Assurance
Sampling), 'lab_specimen' (Human Specimen),
'lab_env' (Environmental Sample), 'lab_specimen_virus'
(Human Specimen Viruses). Default is 'cases'.}

\item{region}{Region code for data filtering, default is 'AFRO'.
This parameter filters the data by the specified WHO region.}

\item{country_code}{ISO3 country code to filter the data. Default is NULL.}

\item{select_vars}{Vector of variables to select, default is NULL (all vars).
This parameter allows for the selection of specific
variables from the API response.}

\item{file_path}{Path for the data and log files, default is a preset
directory. This defines the storage location for data
and logs.}

\item{save_directly}{Boolean indicating whether data should be saved directly.
Default is TRUE. If FALSE, the function returns a
data frame of the aggregated data.}

\item{log_results}{Boolean indicating whether to log update sessions.
Default is TRUE. This controls the logging of updates.}

\item{polis_api_key}{API key for authentication, default from environment.
This is used for accessing and authenticating with
the API.}
}
\value{
If save_directly is FALSE, returns a data frame of the aggregated
data. Otherwise, the data is saved to the specified file path and
nothing is returned.
}
\description{
This function is designed for both the initial retrieval and subsequent
updates of POLIS datasets. It intelligently checks for existing data,
fetches only new data since the last update, integrates this with the
existing data, logs update sessions, and saves the updated dataset to a
specified location. It interacts with the POLIS API and allows filtering
and selection based on various criteria.
}
\examples{
\dontrun{
update_polis_api_data("2021-01-01", "2021-01-31", "cases", "AFRO")
}
}
