% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ona.R
\name{validate_base_url}
\alias{validate_base_url}
\title{Validate and Normalize a Base URL}
\usage{
validate_base_url(base_url)
}
\arguments{
\item{base_url}{A string representing the base URL to be validated.}
}
\value{
A normalized base URL containing only the protocol and the domain.
If the URL is not valid, the function stops and returns an error message.
}
\description{
This function checks if the provided \code{base_url} adheres to a standard URL
format, specifically ensuring it starts with either \code{http} or \code{https} and
contains no path elements.
}
