% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_data}
\alias{validate_data}
\title{Validate Dataset}
\usage{
validate_data(
  data,
  group_var = NULL,
  id_col = NULL,
  geo_name_cols = NULL,
  geo_id_cols = NULL,
  lat_long_cols = NULL,
  date_cols = NULL,
  date_pair_cols = NULL,
  n_groups = 10,
  decreasing = FALSE,
  plots_path = NULL,
  custom_title = NULL,
  save_output = FALSE,
  save_preffix = NULL,
  vheight = 1400,
  vwidth = 1550,
  autoscale_nanoplot = FALSE,
  ...
)
}
\arguments{
\item{data}{A dataframe containing the AFRO data to be validated.}

\item{group_var}{Optional. The column name to group the data by. If NULL,
uses default.}

\item{id_col}{Optional. The column name for the unique identifier. If NULL,
uses default.}

\item{geo_name_cols}{Optional. A vector of column names for geographic names.
If NULL, uses default.}

\item{geo_id_cols}{Optional. A vector of column names for geographic IDs.
If NULL, uses default.}

\item{lat_long_cols}{Optional. A vector of column names for latitude and
longitude. If NULL, uses default.}

\item{date_cols}{Optional. A vector of column names for date fields. If NULL,
uses default.}

\item{date_pair_cols}{Optional. A list of date column pairs for comparison.
If NULL, uses default.}

\item{n_groups}{Integer. The number of groups to display in the summary.
Default is 10.}

\item{decreasing}{Logical. Whether to sort the groups in decreasing order.
Default is FALSE.}

\item{plots_path}{Optional. The file path to save output plots. Required if
save_output is TRUE.}

\item{custom_title}{Optional. A custom title for the output table.}

\item{save_output}{Logical. Whether to save the output as HTML and PNG.
Default is FALSE.}

\item{save_preffix}{Optional. A prefix for the saved output file name.}

\item{vheight}{Integer. The height of the output image in pixels.
Default is 1400.}

\item{vwidth}{Integer. The width of the output image in pixels. Default
is 1550.}

\item{autoscale_nanoplot}{Logical. Whether to autoscale the nanoplot.
Default is FALSE.}

\item{...}{Additional arguments passed to create_summary_by_group function.}
}
\value{
A list containing two elements:
\item{gt_table}{A gt table object with the validation summary}
\item{id_data}{A dataframe with detailed information for each unique
identifier}
}
\description{
This function performs data quality checks on a given dataset and produces
a gt table with metadata
}
\details{
This function performs various data quality checks on AFRO data before it is
sent to POLIS, including checks for missing values, geographic data
consistency, date field validity, and more. It allows for customization of
column names and grouping variables, making it flexible for different data
structures within the AFRO system.

The function will use default parameters based on the specified data type
(AFP or ES) if custom parameters are not provided. It also checks if all
specified columns exist in the dataset before proceeding with the analysis.
}
\examples{
# Assuming afro_data is your dataset and you have the necessary dependencies
# result <- validate_data(afro_data,
#                        group_var = "ReportingYear",
#                        n_groups = 8,
#                        decreasing = FALSE,
#                        plots_path = "/path/to/save/plots",
#                        save_output = TRUE,
#                        vheight = 1500, vwidth = 1600)

# Access the GT table and ID data
# gt_table <- result$gt_table
# id_data <- result$id_data

}
\seealso{
\code{\link[=create_summary_by_group]{create_summary_by_group()}}
}
