% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_polis_snapshots}
\alias{validate_polis_snapshots}
\title{Validate POLIS Data Snapshots}
\usage{
validate_polis_snapshots(
  path,
  data_type = "AFP",
  year_of_interest = 2024,
  pattern = NULL,
  data_type_string = NULL,
  snapshots = 6,
  plots_path,
  custom_title = NULL,
  vheight = 1400,
  vwidth = 1760,
  ...
)
}
\arguments{
\item{path}{Character. Directory path containing POLIS snapshot files.}

\item{data_type}{Character. Type of data ("AFP" or "ES"). Default is "AFP".}

\item{year_of_interest}{Numeric. Year of interest for validation. Default is
2024.}

\item{pattern}{Character. Pattern to match snapshot files. Default is NULL.}

\item{data_type_string}{String of data list name}

\item{snapshots}{Numeric. Number of snapshots to validate. Default is 6.}

\item{plots_path}{Character. Directory to save output plots.}

\item{custom_title}{Character. Custom title for the output. Default is NULL.}

\item{vheight}{Integer. The height of the output image in pixels.
Default is 1400.  Corresponds to the default 6 snapshots.}

\item{vwidth}{Integer. The width of the output image in pixels. Default
is 1760. Corresponds to the default 6 snapshots.}

\item{...}{Additional arguments passed to validate_polis function.}
}
\value{
A list containing validation results and GT table.
}
\description{
This function validates POLIS data snapshots for a specific year and data
type.
}
