% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world}
\alias{world}
\title{World land data}
\format{
An sf object
}
\source{
Natural Earth \url{https://www.naturalearthdata.com/}
}
\usage{
world
}
\description{
A dataset containing high-resolution world countries data.
}
\keyword{datasets}
