% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_geonames.R
\name{calculate_match_stats}
\alias{calculate_match_stats}
\title{Calculate and Report Geo-naming Match Statistics}
\usage{
calculate_match_stats(
  data,
  lookup_data,
  level0 = NULL,
  level1 = NULL,
  level2 = NULL,
  level3 = NULL
)
}
\arguments{
\item{data}{A dataframe containing the target data to be matched.}

\item{lookup_data}{A dataframe serving as the reference for matching.}

\item{level0}{level0 col name (country) in both 'data' and 'lookup_data'.}

\item{level1}{level1 col name (province/state/region) in both 'data' and
'lookup_data'.}

\item{level2}{level2 col name (district) in both 'data' and 'lookup_data'.}

\item{level3}{level3 col name (subdistrict) in both 'data' and 'lookup_data'.}
}
\value{
Invisible NULL. The primary purpose of this function is to reporting
match stats rather than return a value.
}
\description{
Compares entries in a given dataset against a lookup dataset across specified
admn levels (e.g., countries, (province/state/region), districts,
subdistricts) to calculate and report match statistics.
}
\details{
Calculates unique matches across administrative levels, reports
match and mismatch counts.
}
\examples{
# calculate_match_stats(
#      my_data, my_lookup, "country", "province", "district", "subdistrict")

}
\keyword{internal}
