% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ona.R
\name{get_multi_ona_data}
\alias{get_multi_ona_data}
\title{Get Data from ONA for Multiple Forms}
\usage{
get_multi_ona_data(
  base_url = "https://api.whonghub.org",
  form_ids,
  api_token,
  selected_columns = NULL
)
}
\arguments{
\item{base_url}{The base URL for the ONA API; defaults to
'https://api.whonghub.org'.}

\item{form_ids}{A vector containing form id number to identify each form.}

\item{api_token}{A string specifying the API token for ONA.}

\item{selected_columns}{Selected columns to download. Default is NULL.
It's in stringed list like c("year", "form_id")}
}
\value{
A data frame containing the combined data from all specified form
IDs, and includes from_id column.
}
\description{
This function retrieves data for a specified form from the ONA API using a
provided API token and constructs a unique key for each dataset. It returns
the data in a structured format if the request is successful.
}
\examples{
# api_token <- "your_api_token_here"
}
