% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ona.R
\name{prep_ona_data_endpoints}
\alias{prep_ona_data_endpoints}
\title{Fetch ONA API Data}
\usage{
prep_ona_data_endpoints(base_url = "https://api.whonghub.org", api_token)
}
\arguments{
\item{base_url}{The base URL for the ONA API; defaults to
'https://api.whonghub.org'.}

\item{api_token}{API token for authentication.}
}
\value{
Data frame of the API endpoint data.
}
\description{
Retrieves data from the ONA API endpoints to display available forms and
datasets. It initially, uses given API token for authorization and checks the
initial API response status. Then it validates the base URL and constructs
the full API endpoint URL to retrieve a dataframe of all the available for
forms/data.
}
\examples{
# prep_ona_data_endpoints(api_token = "your_api_token_here")
}
