% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dates.R
\name{validate_date}
\alias{validate_date}
\title{Validate Date Column}
\usage{
validate_date(data, date_col)
}
\arguments{
\item{data}{A data frame containing the date column.}

\item{date_col}{The name of the date column.}
}
\value{
The modified data frame with additional columns indicating issues
found during the checks.
Columns added:
- \verb{\{date_col\}_missing}: Indicates if the date is missing.
- \verb{\{date_col\}_non_date}: Indicates if the value is not a valid date.
- \verb{\{date_col\}_invalid}: Indicates if the date is not sensible (e.g.,
not starting with '20').
- \verb{\{date_col\}_future}: Indicates if the date is in the future.
- \verb{\{date_col\}_leap_issue}: Indicates if there are leap year issues.
- \verb{\{date_col\}_format_issue}: Indicates if there are improper
formatting issues.
}
\description{
This function checks and validates a date column in a data frame. It
performs various checks for missing dates, non-date values, sensible dates,
leap year validity, and date formatting.
}
\examples{
data <- data.frame(
  country = c("Rwanda", "Burundi", "Ethiopia", "Zambia", "Zambia",
              "Chad", "Niger", "Angola"),
  date = c(
    "2023-06-15", "2024-07-20", NA, "1999-12-31", "2025-08-22",
    "2020/23/10", "2020-02-29", "2019-02-29"
  )
)

# check whether dates column is valid 
res <-  poliprep::validate_date(data, "date") 


# Check for countries where there is a 
# non-date issue in the date column
res |> 
 dplyr::filter(date_non_date == TRUE) |> 
 dplyr::count(country)

# Check for countries where there is a 
# invalid leap year issue
res |> 
dplyr::filter(date_leap_issue == TRUE) |> 
dplyr::count(country)

# Check for countries where there is a 
# improper formatting of the date  
res |> 
 dplyr::filter(date_format_issue == TRUE) |> 
 dplyr::count(country)
 
}
